refactor(globals): use getBoolean() for boolean OEGlobalsBag settings
Replace untyped get() calls with getBoolean() for all 212 boolean
globals defined in $GLOBALS_METADATA. A Rector rule parses the metadata
via PhpParser AST to derive types automatically. Surrounding expressions
(comparisons to '0'/'1', bool casts, empty() wrappers) are simplified.

A companion PHPStan rule prevents regression by flagging raw get() for
known-typed globals. Both rules derive their knowledge from the source
of truth in library/globals.inc.php.

Also cast bool to int before passing to js_escape()/attr() in several
files to satisfy their string parameter types.

+++ interface/billing/ub04_form.php
@@ -205,7 +205,7 @@ var ub04id = new Array();
 payerid = <?php echo js_escape($payerid ?? ''); ?>;
 pid = <?php echo js_escape($pid);?>;
 encounter = <?php echo js_escape($encounter ?? null);?>;
-isTemplate = <?php echo js_escape(($isAuthorized === true ? $isAuthorized : false)); ?>;
+isTemplate = <?php echo js_escape((int) ($isAuthorized === true ? $isAuthorized : false)); ?>;
 ub04id = <?php echo $ub04id;?>
 
 function adjustForm()
